Add asciidoc checking to configure.ac, make manpage generation optional
Include manpages when we ship a package tarball, and allow them to be
generated by the end user if they want by using the --enable-asciidoc option
to ./configure. This will allow us to maintain manpages in an easier to modify
format while still keeping the make dependencies to a minimum.

Signed-off-by: Dan McGee <dan@archlinux.org>