libalpm/remove.c : revert part of commit to fix remove044 pactest
In 1c9f30b9fab998e2d89dd307a90122618d746cb6

This change should be done AFTER the TODO in recursedeps is implemented.
Then we'll be able to first sort the list, then call recursedeps which keeps
it sorted. For now, the sorting should still be done as the last step.

Signed-off-by: Chantry Xavier <shiningxc@gmail.com>
Signed-off-by: Dan McGee <dan@archlinux.org>