Clean up architecture handling in configure.ac
Remove the need for two variables for the architecture-specific switches
(things like "-march=i686") by combining it into one variable. Also allow
configure to proceed with only a warning if we don't have presets for the
detected architecture- it is kind of stupid to restrict ourselves like we
had been.

Signed-off-by: Dan McGee <dan@archlinux.org>