Ensure that manpages are always distributed and installed
Commit 012f7939784358b02726c169543aa99436439335 was a bit misguided in its
thinking, and resulted in a package built without asciidoc enabled not
installing the manpages to the system on a 'make install' operation. Fix
this behavior by making manpages required in a normal build, and in order to
disable their existence, the '--disable-doc' option must be used.

Hopefully this solves manpage issues for both developers and package
builders while allowing as much flexibility as possible.

Signed-off-by: Dan McGee <dan@archlinux.org>